requirements.txt
asgiref==3.3.1
astroid==2.5.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.1.4
django-jazzmin==2.4.4
django-phone-field==1.8.1
djangorestframework==3.12.2
idna==2.10
isort==5.8.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
Pillow==8.2.0
pylint==2.7.2
pytz==2020.4
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
wrapt==1.12.1